requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-appconf==1.0.6
django-ckeditor==6.7.1
django-compressor==4.5.1
django-js-asset==2.2.0
django-user-agents==0.4.0
django-htmx==1.23.0
idna==3.10
Markdown==3.7
pillow==11.2.1
python-dotenv==1.0.0
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
sqlparse==0.5.1
tzdata==2025.2
urllib3==2.2.3
django-unfold==0.58.0
Unidecode==1.3.8
django-markdownx==4.0.7
python-magic==0.4.27
PyJWT==2.10.0
cryptography==45.0.2
gunicorn==23.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.10
django-tinymce==4.1.0
django-humanize==0.1.2
beautifulsoup4==4.12.3